api/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
whitenoise>=6.6.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pip-chill>=1.0.3